pyproject.toml
[tool.poetry]
name = "url-shortener"
version = "0.0.1"
description = "Url shortener"
authors = ["kill-your-soul <Platonovdi@gmail.com>"]
readme = "README.md"
packages = [{include = "url_shortener"}]

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2.7"
djangorestframework = "^3.14.0"
pre-commit = "^3.5.0"
shortuuid = "^1.0.11"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
